Got <BV> to stop showing up in LMP deps: * culprit is Black formatter adds space after # comment * it doesn't always though because if the formatter raises exception, no format * fixed by chaning `#<BV>` => `# <BV>` but for all tags
NOTE: this might not be the right way since without rerunning all the LMPs, the old source persists in the DB, so the old #<BV> tags won't get rendered right in the React app. Could instead enforce that # <BV> becomes #<BV> even after formatting, or strip out the delimiter before formatting and add it back in after.